# driver-location/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/v1/locations": {
            "post": {
                "description": "Creates a new driver location",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "locations"
                ],
                "summary": "Create a new driver location",
                "parameters": [
                    {
                        "description": "Create location request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateLocationRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.LocationResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/locations/batch": {
            "post": {
                "description": "Creates multiple driver locations at once",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "locations"
                ],
                "summary": "Create driver locations in bulk",
                "parameters": [
                    {
                        "description": "Create bulk location request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateLocationBulkRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.BulkLocationResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/locations/search": {
            "post": {
                "description": "Searches for driver locations based on a GeoJSON point and radius",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "locations"
                ],
                "summary": "Search for driver locations",
                "parameters": [
                    {
                        "description": "Search location request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.SearchLocationRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.SearchLocationResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/health": {
            "get": {
                "description": "Check if the service is healthy",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "health"
                ],
                "summary": "Check if the service is healthy",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.HealthCheckResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.BulkError": {
            "type": "object",
            "properties": {
                "driver_id": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                }
            }
        },
        "dto.BulkLocationResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/dto.BulkResult"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "dto.BulkResult": {
            "type": "object",
            "properties": {
                "errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.BulkError"
                    }
                },
                "failed": {
                    "type": "integer"
                },
                "successful": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "dto.CreateLocationBulkRequest": {
            "type": "object",
            "required": [
                "locations"
            ],
            "properties": {
                "locations": {
                    "type": "array",
                    "maxItems": 1000,
                    "minItems": 1,
                    "items": {
                        "$ref": "#/definitions/dto.CreateLocationRequest"
                    }
                }
            }
        },
        "dto.CreateLocationRequest": {
            "type": "object",
            "required": [
                "driver_id",
                "latitude",
                "longitude"
            ],
            "properties": {
                "driver_id": {
                    "type": "string"
                },
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                }
            }
        },
        "dto.DriverLocation": {
            "type": "object",
            "properties": {
                "distance": {
                    "type": "number"
                },
                "driver_id": {
                    "type": "string"
                },
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                }
            }
        },
        "dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "dto.HealthCheckResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string",
                    "example": "ok"
                }
            }
        },
        "dto.LocationData": {
            "type": "object",
            "properties": {
                "driver_id": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "dto.LocationResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/dto.LocationData"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "dto.SearchLocationData": {
            "type": "object",
            "properties": {
                "drivers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.DriverLocation"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "dto.SearchLocationRequest": {
            "type": "object",
            "required": [
                "latitude",
                "longitude",
                "radius"
            ],
            "properties": {
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                },
                "radius": {
                    "type": "number",
                    "maximum": 50,
                    "minimum": 0.1
                }
            }
        },
        "dto.SearchLocationResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/dto.SearchLocationData"
                },
                "success": {
                    "type": "boolean"
                }
            }
        }
    }
}